Removed "grades" portion of nav bar, consolidated data into "assignments"
