+++ client/src/components/App/App.jsx
@@ -5,7 +5,6 @@ import { LandingPageProvider } from "./LandingPage/LandingPageContext";
 import StudentsPage from "./StudentsPage/StudentsPage";
 import StudentsLanding from "./StudentsPage/StudentsLanding";
 import Assignments from "./NavMenu/Assignments";
-import Grades from "./NavMenu/Grades";
 import Calendar from "./NavMenu/Calendar";
 import Messages from "./NavMenu/Messages";
 import Error from "../Error";
@@ -44,7 +43,6 @@ const App = () => {
                 <Route path="studentHome" element={<StudentsLanding />} />
                 <Route path="Assignments" element={<Assignments />} />
                 <Route path="Calendar" element={<Calendar />} />
-                <Route path="Grades" element={<Grades />} />
                 <Route path="Messages" element={<Messages />} />
               </Route>
               <Route element={<ProtectedRoute />}>
@@ -59,7 +57,6 @@ const App = () => {
                     element={<InstructorAssignments />}
                   />
                   <Route path="Calendar" element={<Calendar />} />
-                  <Route path="Grades" element={<Grades />} />
                   <Route path="Messages" element={<Messages />} />
                   <Route path="Students" element={<Students />} />
                 </Route>
@@ -68,7 +65,6 @@ const App = () => {
                   <Route path="AdminHome" element={<AdminLandingPage />} />
                   <Route path="Assignments" element={<Assignments />} />
                   <Route path="Calendar" element={<Calendar />} />
-                  <Route path="Grades" element={<Grades />} />
                   <Route path="Messages" element={<Messages />} />
                   <Route path="Students" element={<Students />} />
                 </Route>
