+++ client/src/components/App/StudentsPage/StudentsPage.jsx
@@ -176,23 +176,6 @@ const StudentsPage = () => {
                 <h2 className="text-[1.25rem]  text-center">Calendar</h2>
               </div>
             </Link>
-            <Link to="Grades">
-              <div
-                id="gradesContainer"
-                className={
-                  isDarkMode
-                    ? "flex cursor-pointer flex-col gap-[1rem] p-[2rem] hover:bg-ContentBGDM hover:bg-opacity-[0.75]"
-                    : "flex cursor-pointer flex-col gap-[1rem] p-[2rem] hover:bg-[#afc9c2] hover:bg-opacity-[0.75]"
-                }
-              >
-                {isDarkMode ? (
-                  <img src={gradesDM} alt="gradebook icon" />
-                ) : (
-                  <img src={gradesLM} alt="gradebook icon" />
-                )}
-                <h2 className="text-[1.25rem]  text-center">Grades</h2>
-              </div>
-            </Link>
           </div>
           <div className="w-[100%] h-[100%] flex justify-center">
             <div
